pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentctrl"
version = "0.1.1"
description = "The Institutional Governance Layer for AI Agents — intercept, evaluate, decide."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = []
authors = [
    { name = "MoeIntel" },
]
keywords = [
    "ai-agents",
    "governance",
    "policy-engine",
    "authority",
    "risk-scoring",
    "langchain",
    "openai",
    "crewai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://moeintel.com"
Repository = "https://github.com/moeintel/agentctrl"
Issues = "https://github.com/moeintel/agentctrl/issues"

[project.optional-dependencies]
authority-graph = ["networkx>=3.0"]
langchain = ["langchain-core>=0.2"]
openai-agents = ["openai-agents>=0.1"]
crewai = ["crewai>=0.50"]
all = ["networkx>=3.0", "langchain-core>=0.2", "openai-agents>=0.1", "crewai>=0.50"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentctrl"]
